common.js
